did some cleaning + norm
